Organization and some bug fixes
First of all, I figured out what we can do for the changing of panes,
but it's going to be a pane either way. HAHAHAHAHAHAHAH
Matt and me discussed the issue between using visible panes or
individual forms to represent different states in the application.
Unfortunately, our best bet is to do what we have implemented now. Which
means, for ease of work flow, we need to finish as much as we can with
the main menu first, then the individual session menus, and then finally
the game state panel.  Using the dock in parent toggle for each pane, we
can ensure they all take up the entire window. But if we wish to work on
a previous pane, we have to move everything off to the side so we can
edit it in the form design window. It's tedious, but the alternatives
I'm aware of are not what we want.

Changes:
-Exit buttons for both Client and Host should work now.
-Join Game prompts user with specific ip prompt. Submitting without
typing anything in causes it to search under the local ip. (This hasn't
been fully implemented in the Client class. Won't take long.)
-Organized classes into respective folders.

I had a weird issue where some of the game-related classes weren't
showing up in my solution explorer. I had to reimport them. If something
is missing, let me know.