Removed client's manual query prompt
Made some additional tweaks as well. Still trying to figure out how now
only to parse game related information, but also designate whose turn it
is. Maybe from the host, a method is used to target a specific client to
let it know it is its turn? That way we don't have to muddy up the
client with information the host is already aware of.